Change bounding box types from mat4 to dmat4 for double precision support
Update all bounding box related code to use dmat4/std::optional<dmat4>
instead of mat4/std::optional<mat4> to align with the larger PR that
changed SpatialEntity and Camera to use double precision.

Changes:
- zoomoptions.h: boundingBox std::function now returns optional<dmat4>
- boundingbox.h/cpp: all boundingBox and minExtentBoundingBox functions
  return dmat4; internal vec3/float changed to dvec3/double
- camerautils.h/cpp: boundingBox parameters changed to const dmat4&;
  internal computations updated to double precision
- cameratools.h/cpp: calculateFovBounds and perspectiveZoom updated for dmat4;
  camPts/finalPts arrays changed to dvec3
- cameraproperty.h/cpp: getBoundingBox_ member and constructor parameters
  changed to std::function<std::optional<dmat4>()>
- plottinggl axis helpers: getBoundingBox_ and related functions use dmat4;
  axisprocessorhelper calcScaleFactor functions updated to dmat4
- basegl cameracomponent/periodicitygl/instancerenderer: bounding box
  types changed to dmat4
- oit module: all Rasterizer::boundingBox() virtual functions return
  optional<dmat4>; transformrasterization uses dmat4 cast
- tetramesh: boundingBox lambda wraps mat4 result in dmat4 cast
- base/meshcameraalgorithms: use dvec3/dmat4 for computeNearFar call
- python3 bindings: setCameraView overloads updated to dmat4
- boundingbox unit tests: updated to use dmat4 and double literals

+++ include/inviwo/core/datastructures/camera/cameratools.h
@@ -129,7 +129,7 @@ struct IVW_CORE_API FovBounds {
     bool farPlaneClipped;
 };
 
-IVW_CORE_API FovBounds calculateFovBounds(const glm::mat4& boundingBox, const glm::dvec3& lookFrom,
+IVW_CORE_API FovBounds calculateFovBounds(const glm::dmat4& boundingBox, const glm::dvec3& lookFrom,
                                           const glm::dvec3& lookTo, const glm::dvec3& lookUp,
                                           double nearPlane, double farPlane);
 
@@ -162,7 +162,7 @@ void perspectiveZoom(CamType& cam, const ZoomOptions& opts) {
 
         if (opts.bounded == ZoomOptions::Bounded::Yes && opts.boundingBox &&
             !opts.boundingBox()
-                 .transform([&](const mat4& bb) {
+                 .transform([&](const dmat4& bb) {
                      const auto bounds =
                          calculateFovBounds(bb, newFrom, cam.getLookTo(), cam.getLookUp(),
                                             cam.getNearPlaneDist(), cam.getFarPlaneDist());
